+++ lib/src/progressDialog.dart
@@ -214,11 +214,13 @@ class ProgressDialog implements _ProgressDialog {
 
     var output;
     await future.then((data) {
-      if (onProgressFinish != null) onProgressFinish = onProgressFinish?.call(data);
+      if (onProgressFinish != null)
+        onProgressFinish = onProgressFinish?.call(data);
       output = data;
       pDialog.dismiss();
     }).catchError((error) {
-      if (onProgressError != null) onProgressError = onProgressError?.call(error);
+      if (onProgressError != null)
+        onProgressError = onProgressError?.call(error);
       pDialog.dismiss();
     });
 
@@ -269,7 +271,8 @@ class _ProgressDialogWidget extends StatefulWidget {
   }
 }
 
-class _ProgressDialogWidgetState extends State<_ProgressDialogWidget> implements _ProgressDialog {
+class _ProgressDialogWidgetState extends State<_ProgressDialogWidget>
+    implements _ProgressDialog {
   Widget? _title, _message, _loading;
   Color? _backgroundColor;
 
@@ -280,7 +283,8 @@ class _ProgressDialogWidgetState extends State<_ProgressDialogWidget> implements
 
     Widget? title = _title ?? widget.title;
     Widget message = _message ?? (widget.message ?? SizedBox.shrink());
-    Color backgroundColor = _backgroundColor ?? (widget.backgroundColor ?? generalBarrierColor);
+    Color backgroundColor =
+        _backgroundColor ?? (widget.backgroundColor ?? generalBarrierColor);
     Widget loading = (_loading ?? widget.loadingWidget) ??
         Container(
           padding: EdgeInsets.all(10.0),
@@ -327,7 +331,9 @@ class _ProgressDialogWidgetState extends State<_ProgressDialogWidget> implements
               Expanded(
                 child: DefaultTextStyle(
                   child: Semantics(child: message),
-                  style: (widget.dialogStyle?.contentTextStyle ?? dialogTheme.contentTextStyle) ?? (theme.textTheme.subtitle1 ?? TextStyle()),
+                  style: (widget.dialogStyle?.contentTextStyle ??
+                          dialogTheme.contentTextStyle) ??
+                      (theme.textTheme.subtitle1 ?? TextStyle()),
                 ),
               ),
             ],
@@ -503,10 +509,12 @@ class CustomProgressDialog implements _CustomProgressDialog {
     var output;
     try {
       await future.then((data) {
-        if (onProgressFinish != null) onProgressFinish = onProgressFinish?.call(data);
+        if (onProgressFinish != null)
+          onProgressFinish = onProgressFinish?.call(data);
         output = data;
       }).catchError((error) {
-        if (onProgressError != null) onProgressError = onProgressError?.call(error);
+        if (onProgressError != null)
+          onProgressError = onProgressError?.call(error);
       });
     } catch (e) {}
     pDialog.dismiss();
@@ -523,7 +531,8 @@ class _CustomProgressDialogWidget extends StatefulWidget {
   final double? blur;
   final Color? backgroundColor;
   final bool? dismissable;
-  _CustomProgressDialogWidgetState _dialogWidgetState = _CustomProgressDialogWidgetState();
+  _CustomProgressDialogWidgetState _dialogWidgetState =
+      _CustomProgressDialogWidgetState();
 
   _CustomProgressDialogWidget({
     Key? key,
@@ -551,13 +560,16 @@ class _CustomProgressDialogWidget extends StatefulWidget {
   }
 }
 
-class _CustomProgressDialogWidgetState extends State<_CustomProgressDialogWidget> implements _CustomProgressDialog {
+class _CustomProgressDialogWidgetState
+    extends State<_CustomProgressDialogWidget>
+    implements _CustomProgressDialog {
   Widget? _loadingWidget;
   Color? _backgroundColor;
 
   @override
   Widget build(BuildContext context) {
-    Color backgroundColor = _backgroundColor ?? (widget.backgroundColor ?? generalBarrierColor);
+    Color backgroundColor =
+        _backgroundColor ?? (widget.backgroundColor ?? generalBarrierColor);
     Widget loadingWidget = (this._loadingWidget ?? widget.loadingWidget) ??
         Container(
           padding: EdgeInsets.all(10.0),
@@ -577,7 +589,8 @@ class _CustomProgressDialogWidgetState extends State<_CustomProgressDialogWidget
       onDismiss: widget.onDismiss,
       barrierColor: backgroundColor,
       dialog: Padding(
-        padding: MediaQuery.of(context).viewInsets + const EdgeInsets.symmetric(horizontal: 20.0, vertical: 24.0),
+        padding: MediaQuery.of(context).viewInsets +
+            const EdgeInsets.symmetric(horizontal: 20.0, vertical: 24.0),
         child: Center(
           child: loadingWidget,
         ),