+++ lib/src/zoom_widget/zoom_widget.dart
@@ -83,20 +83,27 @@ class _ZoomState extends State<Zoom> with TickerProviderStateMixin {
 
   @override
   void initState() {
-    scaleAnimation = AnimationController(vsync: this, lowerBound: 0.0, upperBound: 1.0, duration: Duration(milliseconds: 250));
+    scaleAnimation = AnimationController(
+        vsync: this,
+        lowerBound: 0.0,
+        upperBound: 1.0,
+        duration: Duration(milliseconds: 250));
     scaleAnimation?.addListener(() {
       setState(() {
         if (doubleTapDown ?? false) {
-          scale = map(scaleAnimation?.value ?? 0, 0.0, 1.0, doubleTapScale, 1.0);
+          scale =
+              map(scaleAnimation?.value ?? 0, 0.0, 1.0, doubleTapScale, 1.0);
         } else {
           scale = map(
               scaleAnimation?.value ?? 0,
               0.0,
               1.0,
               doubleTapScale,
-              ((globalConstraints?.maxHeight ?? 0) > (globalConstraints?.maxWidth ?? 0))
+              ((globalConstraints?.maxHeight ?? 0) >
+                      (globalConstraints?.maxWidth ?? 0))
                   ? (globalConstraints?.maxWidth ?? 0) / (widget.maxZoomWidth)
-                  : (globalConstraints?.maxHeight ?? 0) / (widget.maxZoomHeight));
+                  : (globalConstraints?.maxHeight ?? 0) /
+                      (widget.maxZoomHeight));
         }
 
         scaleProcess(globalConstraints);
@@ -107,7 +114,9 @@ class _ZoomState extends State<Zoom> with TickerProviderStateMixin {
           widget.onScaleUpdate?.call(scale, zoom);
         }
         if (widget.onPositionUpdate != null) {
-          widget.onPositionUpdate?.call(Offset((auxLeft + localLeft + centerLeft + scaleLeft) * -1, (auxTop + localTop + centerTop + scaleTop) * -1));
+          widget.onPositionUpdate?.call(Offset(
+              (auxLeft + localLeft + centerLeft + scaleLeft) * -1,
+              (auxTop + localTop + centerTop + scaleTop) * -1));
         }
 
         endEscale(globalConstraints);
@@ -122,17 +131,28 @@ class _ZoomState extends State<Zoom> with TickerProviderStateMixin {
     super.dispose();
   }
 
-  double map(double x, double inMin, double inMax, double outMin, double outMax) {
+  double map(
+      double x, double inMin, double inMax, double outMin, double outMax) {
     return (x - inMin) * (outMax - outMin) / (inMax - inMin) + outMin;
   }
 
   void scaleFixPosition(constraints) {
-    if (((widget.maxZoomHeight * scale) > constraints.maxHeight) && ((auxTop + localTop + centerTop + scaleTop) + (widget.maxZoomHeight * scale)) < constraints.maxHeight) {
-      localTop += constraints.maxHeight - ((auxTop + localTop + centerTop + scaleTop) + widget.maxZoomHeight * scale);
+    if (((widget.maxZoomHeight * scale) > constraints.maxHeight) &&
+        ((auxTop + localTop + centerTop + scaleTop) +
+                (widget.maxZoomHeight * scale)) <
+            constraints.maxHeight) {
+      localTop += constraints.maxHeight -
+          ((auxTop + localTop + centerTop + scaleTop) +
+              widget.maxZoomHeight * scale);
     }
 
-    if (((widget.maxZoomWidth * scale) > constraints.maxWidth) && ((auxLeft + localLeft + centerLeft + scaleLeft) + (widget.maxZoomWidth * scale)) < constraints.maxWidth) {
-      localLeft += constraints.maxWidth - ((auxLeft + localLeft + centerLeft + scaleLeft) + widget.maxZoomWidth * scale);
+    if (((widget.maxZoomWidth * scale) > constraints.maxWidth) &&
+        ((auxLeft + localLeft + centerLeft + scaleLeft) +
+                (widget.maxZoomWidth * scale)) <
+            constraints.maxWidth) {
+      localLeft += constraints.maxWidth -
+          ((auxLeft + localLeft + centerLeft + scaleLeft) +
+              widget.maxZoomWidth * scale);
     }
 
     if ((widget.maxZoomHeight * scale) < constraints.maxHeight) {
@@ -149,30 +169,49 @@ class _ZoomState extends State<Zoom> with TickerProviderStateMixin {
     } else
       centerLeft = 0.0;
 
-    zoom = map(scale, 1.0, (constraints.maxHeight > constraints.maxWidth) ? constraints.maxWidth / widget.maxZoomWidth : constraints.maxHeight / widget.maxZoomHeight, 1.0, 0.0);
+    zoom = map(
+        scale,
+        1.0,
+        (constraints.maxHeight > constraints.maxWidth)
+            ? constraints.maxWidth / widget.maxZoomWidth
+            : constraints.maxHeight / widget.maxZoomHeight,
+        1.0,
+        0.0);
   }
 
   void scaleProcess(constraints) {
-    Offset currentMidlePoint = Offset(((auxLeft + localLeft + centerLeft) * -1 + midlePoint.dx) * (1 / scale) - localLeft, ((auxTop + localTop + centerTop) * -1 + midlePoint.dy) * (1 / scale));
+    Offset currentMidlePoint = Offset(
+        ((auxLeft + localLeft + centerLeft) * -1 + midlePoint.dx) *
+                (1 / scale) -
+            localLeft,
+        ((auxTop + localTop + centerTop) * -1 + midlePoint.dy) * (1 / scale));
 
     if (currentMidlePoint.dx > relativeMidlePoint.dx) {
-      double preScaleLeft = (currentMidlePoint.dx - relativeMidlePoint.dx) * scale;
+      double preScaleLeft =
+          (currentMidlePoint.dx - relativeMidlePoint.dx) * scale;
       if (auxLeft + localLeft + preScaleLeft < 0) {
         scaleLeft = preScaleLeft;
       }
     } else {
-      double preScaleLeft = (relativeMidlePoint.dx - currentMidlePoint.dx) * -scale;
-      if ((auxLeft + localLeft + preScaleLeft) > -((widget.maxZoomWidth * scale) - constraints.maxWidth * scale)) scaleLeft = preScaleLeft;
+      double preScaleLeft =
+          (relativeMidlePoint.dx - currentMidlePoint.dx) * -scale;
+      if ((auxLeft + localLeft + preScaleLeft) >
+          -((widget.maxZoomWidth * scale) - constraints.maxWidth * scale))
+        scaleLeft = preScaleLeft;
     }
 
     if (currentMidlePoint.dy > relativeMidlePoint.dy) {
-      double preScaleTop = (currentMidlePoint.dy - relativeMidlePoint.dy) * scale;
+      double preScaleTop =
+          (currentMidlePoint.dy - relativeMidlePoint.dy) * scale;
       if (auxTop + localTop + preScaleTop < 0) {
         scaleTop = preScaleTop;
       }
     } else {
-      double preScaleTop = (relativeMidlePoint.dy - currentMidlePoint.dy) * -scale;
-      if ((auxTop + localTop + preScaleTop) > -((widget.maxZoomHeight * scale) - constraints.maxHeight * scale)) scaleTop = preScaleTop;
+      double preScaleTop =
+          (relativeMidlePoint.dy - currentMidlePoint.dy) * -scale;
+      if ((auxTop + localTop + preScaleTop) >
+          -((widget.maxZoomHeight * scale) - constraints.maxHeight * scale))
+        scaleTop = preScaleTop;
     }
   }
 
@@ -199,7 +238,8 @@ class _ZoomState extends State<Zoom> with TickerProviderStateMixin {
     if (widget.centerOnScale) {
       if (portrait ?? false) {
         if (widget.maxZoomHeight * scale < constraints.maxHeight) {
-          centerTop = (constraints.maxHeight - widget.maxZoomHeight * scale) / 2;
+          centerTop =
+              (constraints.maxHeight - widget.maxZoomHeight * scale) / 2;
         }
       } else {
         if (widget.maxZoomWidth * scale < constraints.maxWidth) {
@@ -208,13 +248,15 @@ class _ZoomState extends State<Zoom> with TickerProviderStateMixin {
       }
     }
 
-    if (constraints.maxHeight > constraints.maxWidth && widget.maxZoomWidth * scale < constraints.maxWidth) {
+    if (constraints.maxHeight > constraints.maxWidth &&
+        widget.maxZoomWidth * scale < constraints.maxWidth) {
       setState(() {
         scale = constraints.maxWidth / widget.maxZoomWidth;
       });
     }
 
-    if (constraints.maxWidth > constraints.maxHeight && widget.maxZoomHeight * scale < constraints.maxHeight) {
+    if (constraints.maxWidth > constraints.maxHeight &&
+        widget.maxZoomHeight * scale < constraints.maxHeight) {
       setState(() {
         scale = constraints.maxHeight / widget.maxZoomHeight;
       });
@@ -227,18 +269,28 @@ class _ZoomState extends State<Zoom> with TickerProviderStateMixin {
       builder: (BuildContext context, BoxConstraints constraints) {
         globalConstraints = constraints;
         if (!initOrientation) {
-          scale = map(widget.initZoom, 1.0, 0.0, 1.0, (constraints.maxHeight > constraints.maxWidth) ? constraints.maxWidth / widget.maxZoomWidth : constraints.maxHeight / widget.maxZoomHeight);
+          scale = map(
+              widget.initZoom,
+              1.0,
+              0.0,
+              1.0,
+              (constraints.maxHeight > constraints.maxWidth)
+                  ? constraints.maxWidth / widget.maxZoomWidth
+                  : constraints.maxHeight / widget.maxZoomHeight);
           initOrientation = true;
-          portrait = (constraints.maxHeight > constraints.maxWidth) ? true : false;
+          portrait =
+              (constraints.maxHeight > constraints.maxWidth) ? true : false;
 
           if (widget.centerOnScale) {
             if (portrait ?? false) {
               if (widget.maxZoomHeight * scale < constraints.maxHeight) {
-                centerTop = (constraints.maxHeight - widget.maxZoomHeight * scale) / 2;
+                centerTop =
+                    (constraints.maxHeight - widget.maxZoomHeight * scale) / 2;
               }
             } else {
               if (widget.maxZoomWidth * scale < constraints.maxWidth) {
-                centerLeft = (constraints.maxWidth - widget.maxZoomWidth * scale) / 2;
+                centerLeft =
+                    (constraints.maxWidth - widget.maxZoomWidth * scale) / 2;
               }
             }
           }
@@ -247,16 +299,20 @@ class _ZoomState extends State<Zoom> with TickerProviderStateMixin {
           }
 
           if (widget.onPositionUpdate != null) {
-            widget.onPositionUpdate?.call(Offset((auxLeft + localLeft + centerLeft + scaleLeft) * -1, (auxTop + localTop + centerTop + scaleTop) * -1));
+            widget.onPositionUpdate?.call(Offset(
+                (auxLeft + localLeft + centerLeft + scaleLeft) * -1,
+                (auxTop + localTop + centerTop + scaleTop) * -1));
           }
         }
 
-        if (!(portrait ?? false) && constraints.maxHeight > constraints.maxWidth) {
+        if (!(portrait ?? false) &&
+            constraints.maxHeight > constraints.maxWidth) {
           portrait = true;
           centerTop = 0;
           centerLeft = 0;
           scale = 1.0;
-        } else if ((portrait ?? false) && constraints.maxHeight <= constraints.maxWidth) {
+        } else if ((portrait ?? false) &&
+            constraints.maxHeight <= constraints.maxWidth) {
           portrait = false;
           centerTop = 0;
           centerLeft = 0;
@@ -265,19 +321,31 @@ class _ZoomState extends State<Zoom> with TickerProviderStateMixin {
 
         return RawGestureDetector(
           gestures: {
-            MultiTouchGestureRecognizer: GestureRecognizerFactoryWithHandlers<MultiTouchGestureRecognizer>(
+            MultiTouchGestureRecognizer: GestureRecognizerFactoryWithHandlers<
+                MultiTouchGestureRecognizer>(
               () => MultiTouchGestureRecognizer(),
               (MultiTouchGestureRecognizer instance) {
                 instance.onSingleTap = (point) {
                   if (widget.doubleTapZoom) {
                     midlePoint = point;
-                    relativeMidlePoint = Offset(((auxLeft + localLeft + centerLeft) * -1 + midlePoint.dx) * (1 / scale), ((auxTop + localTop + centerTop) * -1 + midlePoint.dy) * (1 / scale));
+                    relativeMidlePoint = Offset(
+                        ((auxLeft + localLeft + centerLeft) * -1 +
+                                midlePoint.dx) *
+                            (1 / scale),
+                        ((auxTop + localTop + centerTop) * -1 + midlePoint.dy) *
+                            (1 / scale));
                   }
                 };
                 instance.onMultiTap = (firstPoint, secondPoint) {
-                  midlePoint = Offset((firstPoint.dx + secondPoint.dx) / 2.0, (firstPoint.dy + secondPoint.dy) / 2.0);
+                  midlePoint = Offset((firstPoint.dx + secondPoint.dx) / 2.0,
+                      (firstPoint.dy + secondPoint.dy) / 2.0);
 
-                  relativeMidlePoint = Offset(((auxLeft + localLeft + centerLeft) * -1 + midlePoint.dx) * (1 / scale), ((auxTop + localTop + centerTop) * -1 + midlePoint.dy) * (1 / scale));
+                  relativeMidlePoint = Offset(
+                      ((auxLeft + localLeft + centerLeft) * -1 +
+                              midlePoint.dx) *
+                          (1 / scale),
+                      ((auxTop + localTop + centerTop) * -1 + midlePoint.dy) *
+                          (1 / scale));
                 };
               },
             ),
@@ -314,19 +382,26 @@ class _ZoomState extends State<Zoom> with TickerProviderStateMixin {
               setState(() {
                 if (details.scale != 1.0) {
                   if (details.scale > changeScale) {
-                    double preScale = scale + (details.scale - changeScale) / widget.zoomSensibility;
+                    double preScale = scale +
+                        (details.scale - changeScale) / widget.zoomSensibility;
                     if (preScale < 1.0) {
                       scale = preScale;
                     }
-                  } else if (changeScale > details.scale && (widget.maxZoomWidth * scale > constraints.maxWidth || widget.maxZoomHeight * scale > constraints.maxHeight)) {
-                    double preScale = scale - (changeScale - details.scale) / widget.zoomSensibility;
+                  } else if (changeScale > details.scale &&
+                      (widget.maxZoomWidth * scale > constraints.maxWidth ||
+                          widget.maxZoomHeight * scale >
+                              constraints.maxHeight)) {
+                    double preScale = scale -
+                        (changeScale - details.scale) / widget.zoomSensibility;
 
                     if (portrait ?? false) {
-                      if (preScale > (constraints.maxWidth / widget.maxZoomWidth)) {
+                      if (preScale >
+                          (constraints.maxWidth / widget.maxZoomWidth)) {
                         scale = preScale;
                       }
                     } else {
-                      if (preScale > (constraints.maxHeight / widget.maxZoomHeight)) {
+                      if (preScale >
+                          (constraints.maxHeight / widget.maxZoomHeight)) {
                         scale = preScale;
                       }
                     }
@@ -341,21 +416,39 @@ class _ZoomState extends State<Zoom> with TickerProviderStateMixin {
 
                   changeScale = details.scale;
                 } else {
-                  if (details.focalPoint.dy > changeTop && (auxTop + up) < 0 && (auxTop + up) > -((widget.maxZoomHeight) * scale - constraints.maxHeight)) {
+                  if (details.focalPoint.dy > changeTop &&
+                      (auxTop + up) < 0 &&
+                      (auxTop + up) >
+                          -((widget.maxZoomHeight) * scale -
+                              constraints.maxHeight)) {
                     localTop = up;
-                  } else if (changeTop > details.focalPoint.dy && (auxTop + down) < 0 && (auxTop + down) > -((widget.maxZoomHeight) * scale - constraints.maxHeight)) {
+                  } else if (changeTop > details.focalPoint.dy &&
+                      (auxTop + down) < 0 &&
+                      (auxTop + down) >
+                          -((widget.maxZoomHeight) * scale -
+                              constraints.maxHeight)) {
                     localTop = down;
                   }
-                  if (details.focalPoint.dx > changeLeft && (auxLeft + right) < 0 && (auxLeft + right) > -((widget.maxZoomWidth * scale) - constraints.maxWidth)) {
+                  if (details.focalPoint.dx > changeLeft &&
+                      (auxLeft + right) < 0 &&
+                      (auxLeft + right) >
+                          -((widget.maxZoomWidth * scale) -
+                              constraints.maxWidth)) {
                     localLeft = right;
-                  } else if (changeLeft > details.focalPoint.dx && (auxLeft + left) < 0 && (auxLeft + left) > -((widget.maxZoomWidth * scale) - constraints.maxWidth)) {
+                  } else if (changeLeft > details.focalPoint.dx &&
+                      (auxLeft + left) < 0 &&
+                      (auxLeft + left) >
+                          -((widget.maxZoomWidth * scale) -
+                              constraints.maxWidth)) {
                     localLeft = left;
                   }
                 }
               });
 
               if (widget.onPositionUpdate != null) {
-                widget.onPositionUpdate?.call(Offset((auxLeft + localLeft + centerLeft + scaleLeft) * -1, (auxTop + localTop + centerTop + scaleTop) * -1));
+                widget.onPositionUpdate?.call(Offset(
+                    (auxLeft + localLeft + centerLeft + scaleLeft) * -1,
+                    (auxTop + localTop + centerTop + scaleTop) * -1));
               }
             },
             onScaleEnd: (details) {
@@ -374,7 +467,11 @@ class _ZoomState extends State<Zoom> with TickerProviderStateMixin {
                       scale: scale,
                       alignment: Alignment.topLeft,
                       child: Container(
-                        decoration: BoxDecoration(color: widget.canvasColor, boxShadow: widget.canvasShadow != null ? [(widget.canvasShadow ?? BoxShadow())] : null),
+                        decoration: BoxDecoration(
+                            color: widget.canvasColor,
+                            boxShadow: widget.canvasShadow != null
+                                ? [(widget.canvasShadow ?? BoxShadow())]
+                                : null),
                         width: widget.maxZoomWidth,
                         height: widget.maxZoomHeight,
                         child: widget.child,
@@ -383,24 +480,39 @@ class _ZoomState extends State<Zoom> with TickerProviderStateMixin {
                   ),
                   Positioned(
                     top: constraints.maxHeight - widget.scrollWeight,
-                    left: -(auxLeft + localLeft + centerLeft + scaleLeft) / ((widget.maxZoomWidth * scale) / constraints.maxWidth),
+                    left: -(auxLeft + localLeft + centerLeft + scaleLeft) /
+                        ((widget.maxZoomWidth * scale) / constraints.maxWidth),
                     child: Opacity(
-                      opacity: (widget.maxZoomWidth * scale <= constraints.maxWidth || !widget.enableScroll) ? 0 : widget.opacityScrollBars,
+                      opacity: (widget.maxZoomWidth * scale <=
+                                  constraints.maxWidth ||
+                              !widget.enableScroll)
+                          ? 0
+                          : widget.opacityScrollBars,
                       child: Container(
                         height: widget.scrollWeight,
-                        width: constraints.maxWidth / ((widget.maxZoomWidth * scale) / constraints.maxWidth),
+                        width: constraints.maxWidth /
+                            ((widget.maxZoomWidth * scale) /
+                                constraints.maxWidth),
                         color: widget.colorScrollBars,
                       ),
                     ),
                   ),
                   Positioned(
-                    top: -(auxTop + localTop + centerTop + scaleTop) / ((widget.maxZoomHeight * scale) / constraints.maxHeight),
+                    top: -(auxTop + localTop + centerTop + scaleTop) /
+                        ((widget.maxZoomHeight * scale) /
+                            constraints.maxHeight),
                     left: constraints.maxWidth - widget.scrollWeight,
                     child: Opacity(
-                      opacity: (widget.maxZoomHeight * scale <= constraints.maxHeight || !widget.enableScroll) ? 0 : widget.opacityScrollBars,
+                      opacity: (widget.maxZoomHeight * scale <=
+                                  constraints.maxHeight ||
+                              !widget.enableScroll)
+                          ? 0
+                          : widget.opacityScrollBars,
                       child: Container(
                         width: widget.scrollWeight,
-                        height: constraints.maxHeight / ((widget.maxZoomHeight * scale) / constraints.maxHeight),
+                        height: constraints.maxHeight /
+                            ((widget.maxZoomHeight * scale) /
+                                constraints.maxHeight),
                         color: widget.colorScrollBars,
                       ),
                     ),